+++ android/src/main/kotlin/com/baseflow/permissionhandler/PermissionHandlerPlugin.kt
@@ -7,11 +7,12 @@ import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
 import android.os.Build
 import android.provider.Settings
-import androidx.legacy.app.ActivityCompat
+import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import android.util.Log
 import com.baseflow.permissionhandler.data.PermissionGroup
 import com.baseflow.permissionhandler.data.PermissionStatus
+import com.baseflow.permissionhandler.data.ServiceStatus
 import com.baseflow.permissionhandler.utils.Codec
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
@@ -99,7 +100,14 @@ class PermissionHandlerPlugin(private val registrar: Registrar, private var requ
             call.method == "checkPermissionStatus" -> {
                 val permission = Codec.decodePermissionGroup(call.arguments)
                 val permissionStatus = checkPermissionStatus(permission)
-                handleSuccess(permissionStatus, result)
+
+                result?.success(Codec.encodePermissionStatus(permissionStatus))
+            }
+            call.method == "checkServiceStatus" -> {
+                val permission = Codec.decodePermissionGroup(call.arguments)
+                val serviceStatus = checkServiceStatus(permission)
+
+                result?.success(Codec.encodeServiceStatus(serviceStatus))
             }
             call.method == "requestPermissions" -> {
                 if (mResult != null) {
@@ -149,8 +157,13 @@ class PermissionHandlerPlugin(private val registrar: Registrar, private var requ
         val targetsMOrHigher = context.applicationInfo.targetSdkVersion >= android.os.Build.VERSION_CODES.M
 
         for (name in names) {
-            if (targetsMOrHigher && ContextCompat.checkSelfPermission(context, name) != PackageManager.PERMISSION_GRANTED) {
-                return PermissionStatus.DENIED
+            if (targetsMOrHigher) {
+                val permissionStatus = ContextCompat.checkSelfPermission(context, name)
+                if (permissionStatus == PackageManager.PERMISSION_DENIED) {
+                    return PermissionStatus.DENIED
+                } else if (permissionStatus != PackageManager.PERMISSION_GRANTED) {
+                    return PermissionStatus.UNKNOWN
+                }
             }
         }
 
@@ -161,6 +174,20 @@ class PermissionHandlerPlugin(private val registrar: Registrar, private var requ
         }
 
         return PermissionStatus.GRANTED
+    }
+
+    private fun checkServiceStatus(permission: PermissionGroup) : ServiceStatus {
+        val context: Context? = registrar.activity() ?: registrar.activeContext()
+        if (context == null) {
+            Log.d(mLogTag, "Unable to detect current Activity or App Context.")
+            return ServiceStatus.UNKNOWN
+        }
+
+        if (permission == PermissionGroup.LOCATION || permission == PermissionGroup.LOCATION_ALWAYS || permission == PermissionGroup.LOCATION_WHEN_IN_USE) {
+            return if(isLocationServiceEnabled(context)) ServiceStatus.ENABLED else ServiceStatus.DISABLED
+        }
+
+        return ServiceStatus.NOT_APPLICABLE
     }
 
     private fun shouldShowRequestPermissionRationale(permission: PermissionGroup) : Boolean {
@@ -258,7 +285,10 @@ class PermissionHandlerPlugin(private val registrar: Registrar, private var requ
             } else if (permission == PermissionGroup.LOCATION) {
                 val context: Context? = registrar.activity() ?: registrar.activeContext()
                 val isLocationServiceEnabled= if (context == null) false else isLocationServiceEnabled(context)
-                val permissionStatus = if (isLocationServiceEnabled) grantResults[i].toPermissionStatus() else PermissionStatus.DISABLED
+                var permissionStatus = grantResults[i].toPermissionStatus()
+                if (permissionStatus == PermissionStatus.GRANTED && !isLocationServiceEnabled) {
+                    permissionStatus = PermissionStatus.DISABLED
+                }
 
                 if (!mRequestResults.containsKey(PermissionGroup.LOCATION_ALWAYS)) {
                     mRequestResults[PermissionGroup.LOCATION_ALWAYS] = permissionStatus
@@ -470,9 +500,5 @@ class PermissionHandlerPlugin(private val registrar: Registrar, private var requ
             locationProviders = Settings.Secure.getString(context.contentResolver, Settings.Secure.LOCATION_PROVIDERS_ALLOWED)
             return !TextUtils.isEmpty(locationProviders)
         }
-    }
-
-    private fun handleSuccess(permissionStatus: PermissionStatus, result: Result?) {
-        result?.success(Codec.encodePermissionStatus(permissionStatus))
     }
 }