+++ lib/permission_handler.dart
@@ -29,13 +29,25 @@ class PermissionHandler {
 
   final MethodChannel _methodChannel;
 
+  /// Check current permission status.
+  ///
   /// Returns a [Future] containing the current permission status for the supplied [PermissionGroup].
   Future<PermissionStatus> checkPermissionStatus(
       PermissionGroup permission) async {
-    final dynamic status = await _methodChannel.invokeMethod(
+    final String status = await _methodChannel.invokeMethod(
         'checkPermissionStatus', Codec.encodePermissionGroup(permission));
 
     return Codec.decodePermissionStatus(status);
+  }
+
+  /// Check current service status.
+  ///
+  /// Returns a [Future] containing the current service status for the supplied [PermissionGroup].
+  Future<ServiceStatus> checkServiceStatus(PermissionGroup permission) async {
+    final String status = await _methodChannel.invokeMethod(
+        'checkServiceStatus', Codec.encodePermissionGroup(permission));
+
+    return Codec.decodeServiceStatus(status);
   }
 
   /// Open the App settings page.
@@ -43,6 +55,7 @@ class PermissionHandler {
   /// Returns [true] if the app settings page could be opened, otherwise [false] is returned.
   Future<bool> openAppSettings() async {
     final bool hasOpened = await _methodChannel.invokeMethod('openAppSettings');
+
     return hasOpened;
   }
 
@@ -52,7 +65,7 @@ class PermissionHandler {
   Future<Map<PermissionGroup, PermissionStatus>> requestPermissions(
       List<PermissionGroup> permissions) async {
     final String jsonData = Codec.encodePermissionGroups(permissions);
-    final dynamic status =
+    final String status =
         await _methodChannel.invokeMethod('requestPermissions', jsonData);
 
     return Codec.decodePermissionRequestResult(status);